Move getFingerprint() functionality into Certificate model, add tests for getFingerprint()
