+++ data-prepper-core/src/main/java/org/opensearch/dataprepper/peerforwarder/PeerClientPool.java
@@ -5,7 +5,6 @@
 
 package org.opensearch.dataprepper.peerforwarder;
 
-import com.google.common.io.BaseEncoding;
 import com.linecorp.armeria.client.ClientBuilder;
 import com.linecorp.armeria.client.ClientFactory;
 import com.linecorp.armeria.client.ClientFactoryBuilder;
@@ -13,25 +12,16 @@ import com.linecorp.armeria.client.Clients;
 import com.linecorp.armeria.client.WebClient;
 import io.netty.handler.ssl.util.FingerprintTrustManagerFactory;
 import org.opensearch.dataprepper.plugins.certificate.model.Certificate;
-import org.opensearch.dataprepper.peerforwarder.exception.CertificateFingerprintParsingException;
 
 import java.io.ByteArrayInputStream;
 import java.nio.charset.StandardCharsets;
 import java.time.Duration;
-import java.security.NoSuchAlgorithmException;
-import java.security.cert.CertificateEncodingException;
-import java.security.cert.CertificateException;
-import java.security.cert.CertificateFactory;
-import java.security.cert.X509Certificate;
-import java.security.MessageDigest;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 
 public class PeerClientPool {
     private static final String HTTP = "http";
     private static final String HTTPS = "https";
-    private static final String SSL_ALGORITHM = "X.509";
-    private static final String RSA_ALGORITHM = "SHA-1";
     private final Map<String, WebClient> peerClients;
 
     private int port;
@@ -88,7 +78,7 @@ public class PeerClientPool {
             final ClientFactoryBuilder clientFactoryBuilder = ClientFactory.builder();
 
             if (sslFingerprintVerificationOnly) {
-                final FingerprintTrustManagerFactory fingerprintTrustManagerFactory = new FingerprintTrustManagerFactory(getCertificateFingerPrint());
+                final FingerprintTrustManagerFactory fingerprintTrustManagerFactory = new FingerprintTrustManagerFactory(certificate.getFingerprint());
                 clientFactoryBuilder.tlsCustomizer(sslContextBuilder -> sslContextBuilder.trustManager(fingerprintTrustManagerFactory));
             } else {
                 clientFactoryBuilder.tlsCustomizer(sslContextBuilder -> sslContextBuilder.trustManager(
@@ -112,33 +102,5 @@ public class PeerClientPool {
         }
 
         return clientBuilder.build(WebClient.class);
-    }
-
-    private String getCertificateFingerPrint() {
-        final X509Certificate x509Certificate = getX509Certificate();
-        return convertX509CertificateToFingerprint(x509Certificate);
-    }
-
-    private X509Certificate getX509Certificate() {
-        try {
-            final CertificateFactory certificateFactory = CertificateFactory.getInstance(SSL_ALGORITHM);
-            return (X509Certificate) certificateFactory.generateCertificate(
-                    new ByteArrayInputStream(certificate.getCertificate().getBytes(StandardCharsets.UTF_8))
-            );
-        } catch (final CertificateException e) {
-            throw new CertificateFingerprintParsingException("Unable to convert certificate to X509Certificate object", e);
-        }
-    }
-
-    private String convertX509CertificateToFingerprint(final X509Certificate x509Certificate) {
-        try {
-            final MessageDigest messageDigest = MessageDigest.getInstance(RSA_ALGORITHM);
-            final byte[] derEncodedCertificate = x509Certificate.getEncoded();
-            messageDigest.update(derEncodedCertificate);
-            final byte[] digest = messageDigest.digest();
-            return BaseEncoding.base16().lowerCase().encode(digest);
-        } catch (final NoSuchAlgorithmException | CertificateEncodingException e) {
-            throw new CertificateFingerprintParsingException("Unable to convert x509Certificate to hexadecimal fingerprint", e);
-        }
     }
 }
